[gem] Add StandardizeTensorCalculator (#2108)
Summary: Adds a calculator to standardize image tensors along the
channel axis. This is useful for classification models as most
classification models standardize to imagenet means/stddevs. The current
implementation is a bit of a hack to avoid a copy. Should think about
the interface more to see if there's a way to avoid the copy without
being unsafe.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: Tested end to end with the classification pipeline.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 13f5711ec57104b61b88f37b33083c7858d9bcfa