Update linter (#2139)
Summary: The linter go version was too old and handle a patch version in
the go version in go.mod file.
Updating the linter causes issues with pytype which we decided to not
run via megalinter to removing it.
Finally removing a rule that no longer exists in stylelint.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Linter works.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 9fe55813bef37453d977899d8ba25cd8e96539a9