Cap number of detection metrics (#2148)
Summary: As a suggestion from Zain, we should be capping the number of
metrics we send out for our detections. At some point, having data for
so many classes is not useful. This caps the # of metrics to the first
20 classes.

Relevant Issues: towards GML-405

Type of change: /kind cleanup

Test Plan: run GEM and check metrics output

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: f2f2a83bcd765c1d8bd42ade6de3b5f64715f83a