[build] Support python images (#2179)
Summary: Support python images in bazel with a new `gml_py_image` macro.
For now the image is based on `docker.io/python`, instead of a sysroot.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Tested that the image works and that cuda is available in the
container.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: bcdd5f738e123574dc48b712508ec0e2ba6ae483