[build] Faster (cached) python image builds (#2184)
Summary: Adds a `.fast` target to each `gml_py_image` target, that
tradesoff image size for build time. The `.fast` image includes all pip
dependencies in the image so the image is larger but it better utilizes
the cache. Since the pip dependencies are always fixed, that part of the
image doesn't need to get rebuilt with every change to `.py` source
files. Also, the pip dependencies are in a separate layer of the image,
so that layer can be cached on the container runtime side as well. This
takes the build time for the image from about 5 minutes to about 50s
(using the recommended flags I added to DEVELOPMENT.md without the flags
the times are about 10min and about 2.5min).

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Tested that both the `.fast` and the regular image target can
still load CUDA.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 3007af2a5dffae492d9f99bb0260416cf53fb79d