Add a sysroot that includes some extra libs to run an optimized ffmpeg (#2195)
Summary: This created an optional sysroot (not for distribution) that
includes extra libs and python3.10 in the base image. We should be
able to use images based on this sysroot to run ffmpeg with HW accel.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Built a python image with the new sysroot and ensured that it
included the new libs.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 2e6702766a544b8cdd15d7892068a93418a28d85