Fixup python images that use the sysroot (#2215)
Summary: The python images that use the base from docker work.
The sysroot images have the executable in a different location so need
the `PYTHONPATH` set to search for pip installed packages. The sysroot
also needs coreutils since the minimal python image runs a wrapper to
setup the paths.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Built the cuda test image with the sysroot and tested it.
Also tested the gen_captions_image and ensured it works.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: cf31cadf00683e1fe78332a648428c23eb3b540e