[research] Add command to deploy a dev container that rsyncs local files to the container. (#2213)
Summary: Adds a command to the research CLI to deploy a dev container
with any number of GPUs. The container's `/synced` volume will then be
rsynced to the chosen local folder. This allows for changes to quickly
show up in the dev pod without needing redeploys.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Tested that the dev container deploys successfully and that
local changes get reflected in the container's filesystem.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: e29ee396415b0aea7fb5d714b6165b0be5e8c7e6