[research] Add --jupyter to launch_dev command (#2228)
Summary: Add flag to deploy jupyter to dev container. The notebooks will
be stored in your `$USER-dev` folder in experiment data (either in GCS
or the office NFS depending on the cluster). Sets up a port forward for
jupyter to whatever machine you run the cli on, so that you should be
able to just click the link jupyter spits out.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Tested that the jupyter notebook spins up properly and that
the GPU works.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 5e2493bb9b6bf9133201ebe2f370ab6211d9c5f6