[gem] model_exec_handler: handle DeleteExecutionGraph message (#2279)
Summary: Users want to be able to stop a model from running on a device.

The `DeletePhysicalPipeline` mutation and server level implementation
already exist, but the GEM implementation is missing. This PR adds
handling for the `DeleteExecutionGraph` message to the GEM.

Relevant Issues: towards GML-395

Type of change: /kind feature

Test Plan: Deploy the GEM to the NUC. Deploy a model, use gqlsandbox to
delete the pipeline. Check the GEM logs to make sure it works.

Changelog Message:

---------

Signed-off-by: Hannah Troisi <hannah@gimletlabs.ai>
GitOrigin-RevId: c0185aa5240dfd3fa46d5462efc7ad1b3f30e394