Move auth configs to helm values (#2287)
Summary: This moves the auth configs up to the helm values. Also removes
the duplication of the baseURL config and prefixes env vars with GML_

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Deployed a controlplane and ensured that I can still
login/logout and sessions work.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: becb5d519cb4912dbdd4facd038aa154375b2930