[recorded-gem] add VideoHeader to message unpacking (#2338)
Summary: The recorded GEM unpacks messages so that it can log message
counts.
It was missing the recently added VideoHeader type, which caused a
crash. This
PR fixes that.

Relevant Issues: towards GML-435

Type of change: /kind bug

Test Plan: Skaffold the recorded GEM and look at the logs to make sure
it's no
longer crashing due to an unknown type.

Changelog Message:

Signed-off-by: Hannah Troisi <hannah@gimletlabs.ai>
GitOrigin-RevId: 3fc3c70f9bd7dc8a65968b6b20ea17c9d2e3eba2