[fakegem] Update state machine to remove compiling state (#2374)
Summary: This PR updates the fakegem state machine to consist of just
two states: Idle, ModelRunning.

The recent addition of the default pipeline to the GEM is incompatible
with
the original fakegem state machine which expects a `DeviceCapabilities`
message to be sent during the IDLE state. As of now, the
`DeviceCapabilities`
message is sent during the UNKNOWN state, which breaks the GEM.

Relevant Issues: towards GML-435

Type of change: /kind feature

Test Plan: Run a recorded-gem following the instructions
[here](https://github.com/gimletlabs/gimlet/blob/main/src/testutils/README.md#runningreplaying-with-the-fake-gem)
and make sure it doesn't crash and plays video in the UI.

---------

Signed-off-by: Hannah Troisi <hannah@gimletlabs.ai>
GitOrigin-RevId: 8eedd093f3dddd03b32d8911c4f4446b391b2a4f