[crucible] Handle embeddings in coordinator processor (#2389)
Summary: In previous PRs for the coordinator processor, we updated it to
handle frames and bboxes. We held off from handling embeddings/captions
until our decision on pgvector.
With that decision made, this PR implements HandleEmbedding. It first
ensures that the embedding table exists for a particular model, and
inserts the embeddings accordingly.

Relevant Issues: towards GML-411

Type of change: /kind feature

Test Plan: unit test

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: bc8bdb441a7c6fe55cb91b751dd452812fdbf0f6