[linter dockerfile] Fix tflint COPY command to better represent what actually happens (#2425)
Summary: $HOME doesn't get resolved in the COPY command of the
dockerfile, so the file is written to /.tflint.hcl

This happens to work because we run the dependent command at / anyways.
But I'm fixing because this gives the impression that you can copy to
$HOME in the same way, which I ran into when trying to support a python
linter.

Relevant Issues: n/a

Type of change: /kind bugfix

Test Plan: Ensure that /root/.tflint.d still shows up after building.

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: ee9d5c6bde97b7ce1b5a404fb9fb817596b2c3cf