[gem] Pass dispatcher through to metric scrapers (#2503)
Summary: We need to poll for GPU utilization for jetsons. In order to do
that we need to pass the dispatcher down to the metric scrapers. This PR
adds it to the signature of metric scraper builders in the plugin
registry.

Relevant Issues: Towards GML-467.

Type of change: /kind cleanup

Test Plan: Should be minimal changes here, tested that things still
work.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 32d42d2e6f395dc4d765a039c93e7a337782932e