[gem] Poll for GPU utilization on the jetson. (#2508)
Summary: Add polling for GPU utilization in jetson metric scraper. Since
the jetson only provides a file with a value from 0 to 1000 for current
instantaneous util, we need to continuously poll that file to build a
counter for utilization.

Relevant Issues: Fixes GML-467.

Type of change: /kind feature

Test Plan: Tested by deploying to a Jetson, checked that the metrics
looked reasonable.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 15a9c191659b30f83d59557f06dbfe9929680b03