Fix healthz checks (#2570)
Summary: Go 1.22 updated the `net/http` package to support an optional
METHOD in the http handler installation. See
[here](https://pkg.go.dev/net/http@go1.22.0#hdr-Patterns) This meant
that our install of `/healthz/DB Connected` was failing due to the
unescaped space. This fixes the same.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Sakffolded a dev controlplane and verified that it no longer
crashloops and registers healthz checks as expecter.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 126c3f3fc6516cca19cd6f423b571ea493ff5911