Persist current running logical pipeline ID in ControlExecutionContext (#2781)
Summary: The ControlExecutionContext is a nice common interface among
calculators and various message handlers. So it seems reasonable to have
it be the store of the current logical pipeline ID. This is set by the
model exec handler and will be consumed by various metrics.

Relevant Issues: Towards GML-557

Type of change: /kind infra

Test Plan: Logged that the context is being invoked. Need to test with a
controlplane that actually populates this in the proto.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 0549da1024aecfd6d13b1d9808563838dfa74cd6