[FleetMgr] Store fleet_id in deploy_key_tag table (#2955)
Summary: This PR adds `fleet_id` as a column to the `deploy_key_tag`
table and `AssociateTagsWithDeployKeyRequest`. This fleet_id information
will be used by ListTagsAssociatedWithDeployKey (#2908) to verify that
the caller has permissions to the deploy key.

Relevant Issues: n/a

Type of change: /kind feature

Test Plan: Update the unit tests.

Signed-off-by: Hannah Troisi <hannah@gimletlabs.ai>
GitOrigin-RevId: 64e1d9492bbd8e70a44103bc8e6ce0af1bb61b4f