[FleetMgr] Add ListTagsAssociatedWithDeployKey message (#2908)
Summary: We want to display tags in the deploy key table. To do this,
the fleet
resolver needs to be able list all tags associated with a deploy key. So
this PR
adds the `ListTagsAssociatedWithDeployKey` to the FleetMgr.

Relevant Issues: towards GML-653

Type of change: /kind feature

Test Plan: Add a unit test.

Signed-off-by: Hannah Troisi <hannah@gimletlabs.ai>
GitOrigin-RevId: 92202df940ddd39b546a36fc6262d5e29f1485e3