Rollback for failed idempotency transactions (#2978)
Summary: Ran into this issue when cleaning up the fleetmgr tests. This
surfaced as an issue where the fleetmgr test was hanging. What was
actually happening was that the idempotency key already existed, but
since the tx never gets rolled back, we couldn't create a new
connection.
Whenever we're using `CreateIdempotentTx`, if it errors, we immediately
error without rolling back. We should make sure to do the rollback in
`CreateIdempotentTx` so that the caller doesn't have to.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: existing tests should pass, update idempotent tx test to
check the rollback is called

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 28263582504ee7e4eb596500a92db48e16c14516