[pyproject] Move isort config to pyproject.toml and ensure that isort follows black conventions (#3003)
Summary: isort and black were flapping in my vscode editor (see
https://blog.osull.com/2022/03/02/python-vs-code-make-black-and-organize-imports-work-together-on-save/
for video). This is because they were competing to format the imports.

You can configure isort to use black conventions by setting the
configuration in `pyproject.toml` or in `.isort.cfg`. James pointed out
in an earlier version of this PR that we already have a configuration
for isort. I was confused why isort wouldn't listen to this setting, and
it turns out the isort.cfg file was written incorrectly.

Based on
https://pycqa.github.io/isort/docs/configuration/config_files.html
we should have had this in `.isort.cfg
```
[settings]
profile = "black"
```
but we had
```
[tool.isort]
profile = "black"
```
which isort didn't parse.

So an alternative to this PR would be to fix the isort cfg, but
consolidation in pyproject.toml is equally supported and easier to
manage going forward IMO.

Relevant Issues: n/a

Type of change: /kind ide

Test Plan: Tested on my vscode file and stopped seeing flapping

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 6c949695d2a1537f3c5184241a452e66dffd79ac