chore: Use forked lavis to drop spacy from requirements (#3127)
Summary: This switches the lavis dep to come from a forked version that
drops
spacy from the dependencies.
This is in prep to upgrade gradio since spacy and gradio conflict on
what
version of typer they want.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Test that crucible still works.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 8a540555c22ef1871844440199a0916e67605209