chore: Upgrade nextjs (#3156)
Summary: This upgrades nextjs to the latest release version.
Note that nextjs changed the tracer used for standalone builds in
`v13.5.4` (see
https://github.com/vercel/next.js/commit/eba9fca3ffef460c4431c42db0197ae05dc3fbb2)
which seems to have broken tracing in our setup. This isn't unique to us
and there are other issues filed upstream for this bug (eg
https://github.com/vercel/next.js/issues/62591).
To work around this issue, we build outside the sandbox with pnpm
installed deps.

Relevant Issues: Addresses CVE-2024-34351 CVE-2023-44270 and
CVE-2023-4316

Type of change: /kind infra

Test Plan: Deployed a controlplane via skaffold and ensured it works.
Also ensured that the dev ui build works.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 7c642e6e1d1993304c693b2ec6179ee400455362