Add pipeline deployment grpc protos (#3402)
Summary: This is a part of the pipeline management refactor. We are
adding "PipelineDeployments" as a layer between the logical and physical
pipelines. This adds the protos for deploying and fetching that
information.

Relevant Issues: towards GML-896

Type of change: /kind feature

Test Plan: the protos are unused so far

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: c0dc2084bd63da9de3ea85a2510e99f6aa71f9f8