feature: Improve install script to indicate container failure (#3408)
Summary: The install script used to create a docker container and then
exit with a success message.
Instead now we look through the container logs to find the default model
graph running before declaring success.
If such a message isn't found in 120 seconds, we assume the container
failed and print logs.
If the container exits before either of the former happen, we assume
failure.
The script also leave containers behind (by not adding --rm) so that the
logs can be inspected later if needed.

Relevant Issues: Fixes GML-770

Type of change: /kind cleanup

Test Plan: Ran the install script with a bad controlplane addr. Saw the
docker logs show up in my shell and a message saying gimlet failed to
deploy.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 2fd216964e6b36c2a26f2ddda0a4c6e09c0bf891