Create new physical pipeline protos (#3422)
Summary: The meaning of "physical pipeline" is changing with the
refactor of our state management. This creates a new proto to reflect
the updated physical pipeline schema. This is now going in our corepb
protos, since pipelineDeployment is already in there, as opposed to the
service protos.

Relevant Issues: towards GML-914

Type of change: /kind feature

Test Plan: none, these protos are unused currently

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 59da277e0d2cf13dbe171dfc4361080527a4c606