infra: Update dev and training clusters to be in dev-infra (#3429)
Summary: This creates a dev cluster and training cluster in the
dev-infra project since these are managed by the infra team.
The dev project is supposed to be used for things not managed centrally
by the infra team.

This also updates everything necessary to bring up a dev controlplane in
the new project.

Removal of resources in dev will happen as a followup.

Relevant Issues: Towards GML-853

Type of change: /kind infra

Test Plan: Ran terraform and then deployed a dev controlplane to it.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 515fb4487af7bfc33650acf5bbd01a8c7ec7728c