[build] Ensure gRPC uses our version of re2 (#3445)
Summary: Prior to this PR, gRPC was pulling in its own version of re2.
This wasn't an issue previously, but now that I need components of
tensorflow that rely on re2, I end up with duplicate symbol errors.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Tested that the build still works.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 1b6f7ed5707078ca92cabdcac4e676ead76a9289