infra: Update the logical pipeline definition (#3525)
Summary: Realized that the type for node name was incorrect. Also
realized that we need a node to indicate LambdaInputs since those are
also namespaced in the yaml.
This addresses both of those changes.

Relevant Issues: Towards GML-885

Type of change: /kind infra

Test Plan: Will add tests when the protos are being used.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: dead82827395a3ee42c85fa036f0022a9dfb39dc