[api] Add proto definition for storing information about user provided models (#3566)
Summary: Adds a protobuf that stores information about user uploaded
models. Currently, the main information includes: the format the model
is stored in, the file assets associated with the model, the labels for
classes output by the model if applicable, and the format of the
bounding boxes output by the model if applicable.

Relevant Issues: Towards GML-895

Type of change: /kind feature

Test Plan: Just adds the proto definition, no code to test.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: f9d172c807dacee69ab029bba847b9ee788c858b