Revert "[build] Add more dependencies to Makefile (#3635)" (#3641)
Summary: I didn't realize that this would cause `requirements_lock.txt`
to be updated when running `make gazelle` with unrelated changes in the
relevant build files. I think the pain of the update is so bad, its
better for now to have to manually update the lock file if there's a
relevant change in the BUILD file.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: N/A

GitOrigin-RevId: 27342893dae324d195ff11a097553d45a55eab9e