Update partition handler to allow partitioning by generic entities for controlplane messages (#3632)
Summary: We originally built the partition handlers to partition by
device. However, with the addition of the pipeline deployment
reconciliation, we have a case where we would like to partition by fleet
instead. This is to ensure that the messages are processed in order
according to fleet/pipeline deployment. Since the partition logic is the
same, it made more sense to allow the partition handler to take an
entity ID rather than relying on device ID.

Relevant Issues: towards GML-953

Type of change: /kind cleanup

Test Plan: existing tests pass

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 072845c32df5dd83af87908baab11a213b547bdb