Device disconnect should trigger pipeline deployment reconciliation (#3597)
Summary: A pipeline deployment reconciliation should be triggered any
time a new device connects, a device disconnects, or a change is made to
the deployment spec/state. This adds the control flow for triggering a
reconciliation when a device disconnects. However, it is not as simple
as just triggering an event when the egw bridge is stopped. This is
because we expect the bridge to timeout (every 30 mins?). This uses
jetstream's nakWithDelay so that we can check if the device is actually
disconnected after 3 minutes.

Relevant Issues: towards GML-921

Type of change: /kind feature

Test Plan: add unit test

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 03311e30876445b9af6a8cb5ac37a68d285dd9eb