Force go tests to be in separate packages (#3673)
Summary: When working on #3672 I ran into some issues with testsuites
that included some files in `<packagename>` and others in
`<packagename>_test`. This mixing caused different test behavior
depending on when/how you ran the test. I think we should enforce tests
to be in a separate package and attempt to only test the public api from
a package when possible.

This adds a linter to enfore the same and fixes all existing tests.

(In extreme cases we can override the linter).

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: All tests continue to pass.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 7069dde2f4baa22d02ad14618b9e6468bb4ae713