[python api] Support uploading pipelines from pytorch models (#3685)
Summary: Adds `upload_pipeline` to `gml.Client` to upload a logical
pipeline along with a set of pytorch models. To support that this PR
adds `gml.Model` which is a wrapper around a pytorch model with extra
information such as bounding box format, and image preprocessing steps.

Relevant Issues: Towards GML-895.

Type of change: /kind feature

Test Plan: Added a smoke test for the upload pipeline.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 2a968c070e196f7a0612d732265d74c4971a13a0