[gem] Change ImageShapeCPUTensorCalculator to return INTs (#3689)
Summary: Changes the ImageShapeCPUTensorCalculator to use the interface
defined in the new compiler. These changes made it more convienient for
the new compiler to use this op. Not going to merge this until we're
ready to move over.

Relevant Issues: Towards GML-895

Type of change: /kind feature

Test Plan: Updated the test.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 63483aadb707254dcfece63ad7d352f6d8a949cf