[gem] Ignore detections with empty labels (#3746)
Summary: This ignores detections where the label is the empty string.
This is useful for models that have a default "nothing detected" class
where the label is an empty string.

Relevant Issues: Towards GML-977

Type of change: /kind cleanup

Test Plan: Tested that empty string labels are no longer counted as
detections.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 5832f811156eabe023472ad57d4d888861ee5d6d