[compiler] Add protos for specifying input/output TensorSemantics (#3740)
Summary: Add submessages to the ModelInfo proto to specify the expected
semantics of the input tensors to the model, and the semantics of the
tensors the model outputs. This will be used by the compiler to generate
the necessary pre/post processing steps.

Relevant Issues: Towards GML-976.

Type of change: /kind feature

Test Plan: just adding protos, no tests.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 05f9815f4649ac6e7a5f7cd6671f97bde6c63aa7