Update GEM model exec to handle new pipeline messages (#3745)
Summary: We changed the protos for pipeline execution on the GEMs. This
updates the GEM to use the new protos. This will not be landed until all
changes on the control plane are in and switched over to use the new
pipeline execution.

Relevant Issues: towards GML-626

Type of change: /kind cleanup

Test Plan: bazel run GEM and check it starts a pipeline correctly

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: b9b65ceb45733a51d4631cd0a3ceaf15adc252f8