[compiler] Support `pipeline` dialect `SegmentOp`. (#3779)
Summary: Adds support for a new `SegmentOp` in the pipeline dialect.
This op is converted into image preprocessing, a call to a segmentation
model, mask postprocessing, and conversion to a segmentation proto.

Relevant Issues: Towards GML-978.

Type of change: /kind feature

Test Plan: Added lit test for the conversion. Tested end to end that the
segment op works with an ADAS segmentation model.

---------

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 2e188acbf8356a889cdfc685e6a43e5c028cda26