[build] Eigen: switch to tensorflow's copy (#3803)
Summary: We had two copies of Eigen in our repo, one from TensorFlow,
and one we imported directly. This was okay because Eigen is a
header-only library. Nevertheless, it's better not to have duplicates,
so this PR removes our copy and uses the TensorFlow eigen.

Relevant Issues: Towards GML-982

Type of change: /kind cleanup

Test Plan: Checked that bytetrack still builds properly.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: 048a35cbc69a2cb4261bae1e6df32962daec804d