[gem] bytetrack: maintain class id/label (#3814)
Summary: The out-of-the-box bytetracker was not maintaining class ids.
This PR creates a gimletlabs fork of the repo and modifies it so that
class id is maintained during tracking. The calculator is
correspondingly updated to use the tracked class ids.

Relevant Issues: Towards GML-991

Type of change: /kind feature

Test Plan: Updated the test to check for the tracked label.

Changelog Message: Object tracking now maintains the class label.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: a9592d0b58adb391ab0f51caa54c2b1e87492a8f