[research] Extract git merge-base and diff from merge-base into genfiles for experiment containers (#3839)
Summary: We built features for python experiments that would save the
git merge-base and the diff from that merge-base in the experiments
directory to help improve reproducibility.
However these failed to work in job environments because the git repo
would rightfully not be added in. To get around this, we now capture the
merge-base and the diff from that base
in bazel, expose as a genrule, and then can add them to the images that
we create.

The experiments.py file now reads that data first if the envvar pointing
to it is set.

Relevant Issues: Closes GML-794

Type of change: /kind feature

Test Plan: Added tests for the experiments.py part. Ran this in pegboard
and saw the correct patch and merge base being added.

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: ab0a7fead7d33a0c9f30037b7b1db4f5c83da167