Add hostnetwork as a flag to install.sh (#3887)
Summary: Hostnetwork is not needed except for when connecting to a dev
controlplane. This updates our script so that you can specify it as a
flag. This also adds pid=host to the docker flags, since this will be
used when getting our network stats.

Relevant Issues: towards GML-1022

Type of change: /kind cleanup

Test Plan: run install.sh with/without flags

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 1422b199e22891047a6892a2c2d07658d73ebc06