Add event analytics to the UI (#3897)
Summary: This adds event tracking to the UI. We use the context/provider
pattern to load a singleton analytics object as close to the root of the
app as possible and then access the object and track events at the leaf
level.

Relevant Issues: Towards GML-1020

Type of change: /kind feature

Test Plan: Deployed via skaffold with analytics enabled and checked that
events were sent.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 3de70e4b7b183c0b97598b3c02f0bdaba1d2a447