[python API] Support score segmentation masks (#3949)
Summary: Support segmentation masks where the output is a probability of
each class in the python API. This is not yet supported in the backend,
but PRs to come for that.

Relevant Issues: Towards GML-1075

Type of change: /kind feature

Test Plan: Tested end to end with other PRs that a score based
segmentation mask works.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 8c3c5fa275cf76427f171ae933668139496fdb6b