Rename video file flag (#3983)
Summary: The flag to override the camera and use a file instead was
awkwardly named. Given that this deployment scenario is becoming more
common, rename the flag to make it more ergonomic to use.

Relevant Issues: Towards GML-1030

Type of change: /kind cleanup

Test Plan: Pure rename, will test it with a new gem release.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: fb268ded4786c3ef299f75e084d36d1a1739bece