Improve detection of static video file vs live video stream (#3988)
Summary: One of the RTSP cameras I was testing with has a large negative
number as the frame count instead of -1.
This caused us to treat it as a video file and set an impossible large
sleep time for the next frame.
Relaxing this check fixes this issue.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Tested with a RTSP video and some ffmpeg patches.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 11c59484edc8b84641cb3a5bf362dd3e631a433d