[compiler] Add `Regress` pipeline node (#3994)
Summary: Adds a pipeline node for regression models. This node supports
models that output a numeric predicted value (for example, the weight of
a pig). This PR adds support for converting this node to mediapipe.

Relevant Issues: Towards GML-1077.

Type of change: /kind feature

Test Plan: Added a lit test for the conversion from this node to
mediapipe.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: d762ef4241ec9c70a0d9b802e9b3ac9f4c7beeae