infra: Switch to using sentry's logrus hook (#4039)
Summary: The sentryhook we have doesn't send the stacktrace for the
error to sentry. Instead this switches to using the hook provided by
sentry which sends stacks successfully as of v0.28.0 of the integration.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Created a new error object and logged it with logrus. Saw the
traces populate in sentry.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 6290f0e1f1f04487cbbc118c5e0f8d40315cb617