[lint] Fix bugbear issues in crucible (#4044)
Summary: I want to enable bugbear in our linter:
https://github.com/PyCQA/flake8-bugbear. This helps prevents common bugs
in python where variables are not bound correctly or are updated in
unexpected places.

These are the issues that needed to be fixed in the crucible directory.
The changes were roughly the following
1. Remove unused enumerate() calls
2. Raise 'from' for raise calls nested within except calls

Relevant Issues: n/a

Type of change: /kind cleanup

Test Plan: N/A

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: cb6ce00508912323bddf66c599ea9f9592df51e8