chore: Upgrade an indirect graphql-config dep (#4068)
Summary: We already pull in a newer version of `graphql-config` via
another dep. The linter plugin hasn't been updated in a while but work
with the new version of the dep, so upgrade it to resolve a open CVE.

Relevant Issues: CVE-2024-37890
Fixes GML-1108

Type of change: /kind cve

Test Plan: Ran linter after upgrade.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: a7e13158104482d24facc75bcb1a52ce629d667a