[gem] install.sh: Move all GML options together for readability (#4106)
Summary: Move the key script flags to live together higher in the file
(kind of like flags that we have in other code). This arguably helps
readability, especially if someone is looking to see how they can change
the behavior of the script with environment variables.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Ran install.sh

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
GitOrigin-RevId: b3841341641b21dae5c4b578c551b129e1d81d6e