Update the default tag in the install script (#4131)
Summary: `dev-latest` used to be a volatile tag that would be updated by
any dev build of the GEM image. We now have a `latest` tag that points
to a released GEM so use that instead.
The image location was also accidentally changed in #4106 so this
reverts that change.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Will test after a new GEM is tagged.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 1dfe0fdd986d9678588bbfc0624d04f291efd2a9