Update version stamping (#4130)
Summary: I noticed that since we skaffold the controlplane release and
include the timestamp in the version, each controlplane service ends up
with a slightly different version. This makes looking at sentry
challenging. This change drops the volatile timesstamp from the sentry
reporting (though it's still stamped into the build).
I also noticed that the build_number was missing so this change ensures
that it's set in the workflow.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Will test with a controlplane rc.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 571010894caaa4c46d734c633b50b2846ba1528c