[research] Create experimental sysroot (#4051)
Summary: The experimental image had a dependency on libimage. We used
the python base image which was hard to update. To make this a lot
easier, we can just create a new sysroot for the experimental image
which will make it easier to add debian packages as dependencies.
Ended up creating an extension of the CUDA sysroot by making a separate
function and calling that function in the cuda sysroot and the new
experimental_sysroot.

Relevant Issues: n/a

Type of change: /kind feature

Test Plan: Deployed the experimental image and verify that it trains a
model.

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 02628a0d0b04de2bcc037a251a775e6571044a7c