Fix bug where stopping pipeline always results in "FAILED" state (#4171)
Summary: We are always reporting a pipeline as failed when it is
terminated because stopping the model results in a generic
`CalculatorGraph::Run() failed:` error.
This only reports the status as failed if the stop signal was not
called, but the model has not stopped running. I tried to dive deeper
into the runner to see if we can just avoid erroring on stop, but that
seems more involved on the mediapipe level.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: run GEM and stop the pipeline

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: e0f28916509d7d795031965e26a362708414e4ab