[gem] Record metrics for detection box area and aspect ratio (#4201)
Summary: Record the aspect ratio and the area of boxes as histograms.
The bounds I use are different from what we've used before in our
metrics and are a bit arbitrary, but I'll try to provide the basis for
them below.

For area, many of our boxes will be small. For a 1920x1080 image, a
10x10 box will have an area of ~5e-6, which I somewhat arbitrarily chose
as the second lowest bound. I then tried to capture each order of
magnitude up (1e-5) and the middle point (5e-5). We don't really need
much resolution in the 0.1+ range, as we would consider any box in that
range "large" anyways.

Aspect ratio has a different characteristic because it is a ratio. I got
bounds by simulating random box w/h , sampled uniformly from a range of
5 to 1920 and calculated the distribution stats:
```
Median aspect ratio: 1.00
01th percentile: 0.02
10th percentile: 0.20
25th percentile: 0.50
75th percentile: 2.01
90th percentile: 4.80
99th percentile: 35.91
```
I used this to get some basis for the bounds I chose, but I chose the
cleaner 5 and 50 bounds instead of 4.8 and 35.91 to be the inverse of
10th percentile and 01th percentile.

Relevant Issues: Closes GML-1188

Type of change: /kind feature

Test Plan: Added a test, deployed skaffold and ensured that I saw the
metrics in the GraphQL sandbox
```json
{
  "input": {
    "query": "sum by (class) ( rate(gml_gem_model_box_area_sum{device_id=\"7e73ef3b-b7ae-4505-8181-e39fb288da8c\"}[5m])) / sum by (class) ( rate(gml_gem_model_box_area_count{device_id=\"7e73ef3b-b7ae-4505-8181-e39fb288da8c\"}[5m]))",
    "relative": "15m",
    "step": "15s"
  }
}
```
```gql
query ExampleQuery($input: MetricsQueryInput!) {
  metrics(input: $input) {
    results {
      labels {
        key
        value
      }
      metricSample {
        values
      }
      name
    }
  }
}

```

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 3d45e156fde5ee398f1decf6a1ad96ec738be3f3