[gem] Add ToString() utilities for otel metrics for use in tests (#4221)
Summary: We're going to write a lot of different otel metrics data. We
want to be able to debug them in the tests, which is much easier if they
pretty print. This PR adds the utilities into the testing directory.

Relevant Issues: n/a

Type of change: /kind testing-utils

Test Plan: Added tests and used in a histogram test that I purposefully
failed, to ensure that the operator<<() override worked.

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 980bfc87dd49808a54864a7915f0af9c5ea14b02