[gem/compiler] Switch tracker metadata to explicit proto from vector of removed ids (#4254)
Summary: Switch to explicit proto of vector of removed IDs to simplify
the message passing and future proof the interface, making it easier to
extend.
Made the changes in the compiler and in the mediapipe calculators, which
is fine because we have not shipped either yet out to customers.

Relevant Issues: Towards GML-1190

Type of change: /kind refactor

Test Plan: Updated the tests

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 387f7786a7a4cd78bd90139c97530076a611829e